# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurants API",
        "description": "REST API for the Restaurants SaaS platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current environment"
        }
    ],
    "paths": {
...
        "/api/v1/organization": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Get the active organization",
                "description": "Return the active organization for the authenticated user.",
                "operationId": "organizationShow",
                "responses": {
                    "200": {
                        "description": "The active organization",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "organization": {
                                                    "$ref": "#/components/schemas/Organization"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "The user has no organization"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Organization"
                ],
                "summary": "Update the active organization",
                "description": "Update the active organization for the authenticated user.",
                "operationId": "organizationUpdate",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Grupo Exemplo"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "grupo-exemplo"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organization updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Organization updated successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "organization": {
                                                    "$ref": "#/components/schemas/Organization"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "The user is not allowed to update the organization"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Organization": {
                "required": [
                    "id",
                    "name",
                    "slug",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Grupo Exemplo"
                    },
                    "slug": {
                        "type": "string",
                        "example": "grupo-exemplo"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "description": "Laravel session cookie. It is created automatically after a successful login and sent by the browser on subsequent requests.",
                "name": "laravel-session",
                "in": "cookie"
            }
        }
    },
...
}